requirements.txt
PyBluez==0.23
tqdm==4.59.0
Pillow==8.1.2
